requirements.txt
google-cloud-bigquery==3.25.0
pandas==2.2.3
requests==2.32.3
openpyxl==3.1.5
xlrd==2.0.1
pyarrow==17.0.0
db-dtypes==1.3.1
